# portal-examenes-netec/specs/05-api/openapi.yaml
openapi: 3.1.0
info: {title: Portal de Registro de Exámenes NETEC API, version: 0.2.0-spec}
servers:
  - url: /api/v1
    description: Ruta relativa same-origin; NGINX conserva `/api` al reenviar.
security: [{bearerAuth: []}]
paths:
...
  /exam-requests:
    post:
      operationId: createDraft
      summary: Crea borrador propio; requester y datos derivados se ignoran si llegan del cliente.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ExamRequestInput'}
      responses: {'201': {$ref: '#/components/responses/ExamRequest'}, '400': {$ref: '#/components/responses/Problem'}}
    get:
      operationId: listOwnRequests
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: folio, in: query, schema: {type: string}}
        - {name: participant, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/Status'}}
        - {name: technologyId, in: query, schema: {type: string, format: uuid}}
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      responses: {'200': {$ref: '#/components/responses/Page'}, '400': {$ref: '#/components/responses/Problem'}}
  /exam-requests/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: getOwnRequest
      responses: {'200': {$ref: '#/components/responses/ExamRequest'}, '404': {$ref: '#/components/responses/Problem'}}
    put:
      operationId: updateDraft
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody: {$ref: '#/components/requestBodies/ExamRequestInput'}
      responses: {'200': {$ref: '#/components/responses/ExamRequest'}, '409': {$ref: '#/components/responses/Problem'}}
...
  /catalogs/exams/{id}:
    get:
      operationId: getExam
      parameters: [{$ref: '#/components/parameters/Id'}]
      description: Devuelve detalle y precio aplicable cuando puede resolverse inequívocamente; si faltan dimensiones usa enlace al endpoint price.
      responses: {'200': {description: Examen, content: {application/json: {schema: {$ref: '#/components/schemas/ExamDetail'}}}}, '404': {$ref: '#/components/responses/Problem'}}
...
  /exam-catalog:
    get:
      operationId: searchExamCatalog
      summary: Consulta paginada del catálogo persistido.
      parameters:
        - {name: vendor, in: query, schema: {type: string, maxLength: 120}}
        - {name: code, in: query, schema: {type: string, maxLength: 100}}
        - {name: query, in: query, schema: {type: string, maxLength: 250}}
        - {name: active, in: query, schema: {type: boolean, default: true}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
      responses:
        '200': {description: Página de catálogo, content: {application/json: {schema: {$ref: '#/components/schemas/ExamCatalogPage'}}}}
...
  /exam-catalog/{id}:
    get:
      operationId: getExamCatalogItem
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: Detalle público, content: {application/json: {schema: {$ref: '#/components/schemas/ExamCatalogItem'}}}}
        '404': {$ref: '#/components/responses/Problem'}
...
  /admin/exams/{id}:
    put:
      operationId: updateExam
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IfMatch'}]
      summary: Modifica examen, retake y comentarios con auditoría.
      responses: {'200': {description: Actualizado}, '409': {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    entraOAuth2:
      type: oauth2
      description: Microsoft Entra ID; tenant/cliente se obtienen de configuración por ambiente.
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/token
          scopes:
            ExamRequests.Access: Acceder al Portal de Registro de Exámenes.
...
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 16, maxLength: 128}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string}}
  requestBodies:
    ExamRequestInput:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/ExamRequestInput'}}}
  schemas:
...
    Status: {type: string, enum: [BORRADOR, REGISTRADA, PENDIENTE_NOTIFICACION, ENVIADA_A_APROBADOR, CANCELADA]}
    ParticipantInput:
      type: object
      additionalProperties: false
      required: [firstName, lastName, email]
      properties: {id: {type: string, format: uuid}, firstName: {type: string, maxLength: 100}, lastName: {type: string, maxLength: 100}, secondLastName: {type: [string, 'null'], maxLength: 100}, email: {type: string, format: email, maxLength: 254}, company: {type: [string, 'null'], maxLength: 160}, country: {type: [string, 'null'], pattern: '^[A-Z]{2}$'}, city: {type: [string, 'null'], maxLength: 120}, externalParticipantId: {type: [string, 'null'], maxLength: 80}}
    ExamAssignmentInput:
      type: object
      additionalProperties: false
      required: [examCatalogId, participantIds]
      description: Cantidad, costos, moneda y snapshots son resueltos por backend.
      properties: {examCatalogId: {type: string, format: uuid}, participantIds: {type: array, minItems: 1, maxItems: 100, uniqueItems: true, items: {type: string, format: uuid}}}
    ExamRequestInput:
      type: object
      additionalProperties: false
      description: No admite requester, asesor comercial, destinatario, snapshots, costos, precio de venta, moneda, quantity, lineTotal ni totalAmount. Requester y salesAdvisor se resuelven desde el principal autenticado; el backend resuelve costo base USD. `siteId` es la sede operativa y `organizationalLocationId` queda como alias deprecado.
      properties: {scheduledCourseCode: {type: [string, 'null'], maxLength: 80}, courseTypeId: {type: [string, 'null'], format: uuid}, segmentId: {type: [string, 'null'], format: uuid}, siteId: {type: [string, 'null'], format: uuid}, organizationalLocationId: {type: [string, 'null'], format: uuid, deprecated: true}, companyName: {type: [string, 'null'], minLength: 2, maxLength: 150, description: Texto libre MVP; opcional al guardar borrador y obligatorio al enviar.}, companyId: {type: [string, 'null'], format: uuid, deprecated: true, description: Sin uso en MVP.}, billingReference: {type: [string, 'null'], maxLength: 250}, participants: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ParticipantInput'}}, examAssignments: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ExamAssignmentInput'}}, observations: {type: [string, 'null'], maxLength: 2000}}
...
    Money: {type: object, required: [amount, currency], properties: {amount: {type: number, format: decimal, minimum: 0, multipleOf: 0.0001}, currency: {type: string, pattern: '^[A-Z]{3}$'}, sourceAmount: {type: [number, 'null'], format: decimal}, sourceCurrency: {type: [string, 'null'], pattern: '^[A-Z]{3}$'}, exchangeRate: {type: [number, 'null'], format: decimal}, exchangeRateDate: {type: [string, 'null'], format: date-time}, validFrom: {type: [string, 'null'], format: date}, validTo: {type: [string, 'null'], format: date}}}
    ExamCatalogItem:
      type: object
      additionalProperties: false
      required: [id, vendor, courseName, examCode, examName, baseCost, baseCurrency, active]
      properties:
        id: {type: string, format: uuid}
        vendor: {type: string}
        courseName: {type: string}
        examCode: {type: string}
        examName: {type: string}
        retake: {type: [string, 'null']}
        baseCost: {type: number, format: decimal, minimum: 0, multipleOf: 0.0001}
        baseCurrency: {type: string, const: USD}
        comments: {type: [string, 'null']}
        active: {type: boolean}
    ExamCatalogPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/ExamCatalogItem'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer}
        totalPages: {type: integer}
    ExamDetail: {type: object, required: [id, code, name, vendor, technology, active], properties: {id: {type: string, format: uuid}, code: {type: string}, name: {type: string}, vendor: {type: object}, technology: {type: object}, certification: {type: [object, 'null']}, active: {type: boolean}, price: {anyOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}}}
    RequesterSnapshot: {type: object, readOnly: true, required: [id, name, email], properties: {id: {type: string}, name: {type: string}, email: {type: string, format: email}, area: {type: [string, 'null']}, businessUnit: {type: [string, 'null']}}}
    SalesAdvisorSnapshot: {type: object, readOnly: true, required: [id, name, userPrincipalName], properties: {id: {type: string}, name: {type: string}, userPrincipalName: {type: string, format: email}}}
    ExamRequest: {allOf: [{$ref: '#/components/schemas/ExamRequestInput'}, {type: object, required: [id, status, requester, salesAdvisor, createdAt, version], properties: {id: {type: string, format: uuid}, folio: {type: [string, 'null']}, status: {$ref: '#/components/schemas/Status'}, requester: {$ref: '#/components/schemas/RequesterSnapshot'}, salesAdvisor: {$ref: '#/components/schemas/SalesAdvisorSnapshot'}, companyNameSnapshot: {type: [string, 'null'], maxLength: 150}, site: {type: [object, 'null']}, approvalRecipientName: {type: [string, 'null']}, approvalNotificationStatus: {type: [string, 'null'], enum: [PENDING, SENT, FAILED, DEAD_LETTER, null]}, itemSnapshots: {type: array, items: {type: object}}, totalAmount: {anyOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}, createdAt: {type: string, format: date-time}, submittedAt: {type: [string, 'null'], format: date-time}, version: {type: integer}}}]}
    Problem: {type: object, required: [type, title, status, code, timestamp, correlationId], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}, code: {type: string}, timestamp: {type: string, format: date-time}, correlationId: {type: string}, fieldErrors: {type: array, items: {type: object}}}}
  responses:
    ExamRequest: {description: Solicitud propia, content: {application/json: {schema: {$ref: '#/components/schemas/ExamRequest'}}}}
    Page: {description: Página propia ordenada por fecha descendente, content: {application/json: {schema: {type: object}}}}
...
    Problem: {description: Problem Details, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}